Use --no-startup-id properly
The --no-startup-id parameter disables startup-notification support for
this particular exec command. With startup-notification, i3 can make
sure that a window appears on the workspace on which you used the exec
command. Also, it will change the X11 cursor to watch (a clock) while
the application is launching. So, if an application is not
startup-notification aware (most GTK and Qt using applications seem to
be, though), you will end up with a watch cursor for 60 seconds.