Set up git pull to rebase instead of merge for new branches
When a new branch is created with git branch or git checkout that tracks
another branch, this variable tells Git to set up pull to rebase instead
of merge (see "branch.<name>.rebase"). When never, rebase is never
automatically set to true. When local, rebase is set to true for tracked
branches of other local branches. When remote, rebase is set to true for
tracked branches of remote-tracking branches. When always, rebase will
be set to true for all tracking branches. See "branch.autoSetupMerge"
for details on how to set up a branch to track another branch. This
option defaults to never.